quick-service-sash/services/backend/browse/requirements.txt
flask==3.1.0
elasticsearch==8.16.0
openai==1.57.0
python-dotenv==1.0.1
